requirements.txt
flask
werkzeug
google-genai
python-docx
pypdf
gunicorn
